Move shared test utilities to separate project
