+++ api-test-util/api-test-util.gradle.kts
@@ -1,0 +1,9 @@
+description = "Shared test utilities for the API and app projects."
+
+plugins {
+    `java`
+}
+
+dependencies {
+    compile(project(":api"))
+}
