+++ api/api.gradle.kts
@@ -15,4 +15,5 @@ dependencies {
     api(group = "eu.hansolo", name = "Medusa", version = "7.9") // Note the capital 'M' -- lowercase is a much older version!
     api(group = "com.jfoenix", name = "jfoenix", version = "9.0.8")
     api(group = "com.github.zafarkhaja", name = "java-semver", version = "0.9.0")
+    testCompile(project(":api-test-util"))
 }
