+++ settings.gradle
@@ -1,4 +1,5 @@
 include ":api"
+include ":api-test-util"
 include ":app"
 include ":app:test_plugins"
 include "example-plugins"
